refactor: project 및 version 모듈의 repository/service 구조 정리
